pci: introduce light weight PCIE Host emulation pci-lite
A minimal emulation for PCIE host bridge that supports 0xcf8/0xcfc and
MMFG. Actually there is already a gpex there, but it is designed mainly
for ARM and is not quite suitable for x86:
- it lacks things like PCI hole properties which are required by ACPI.
- the corresponding driver in Linux is designed to work with device tree
  which is not suitable for x86. For this case, additional guest driver
  is even not needed.

Currently MMFG size is limited to 1M, which means only 1 bus is
supported, this is aimed to reduce the scan time in guest. And it
doesn't have a valid vendor ID/device ID assigned so guest may not
recognize it, the functionality however is expected to be OK.

Updated and fixed for newer QEMU by Rob Bradford.

Signed-off-by: Chao Peng <chao.p.peng@linux.intel.com>
Signed-off-by: Rob Bradford <robert.bradford@intel.com>